+++ src/synced_memory/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.1.2"
+__version__ = "0.1.3"
 
 from synced_memory.common import (
     MemoryBase,